requirements.txt
aniso8601==8.1.0
appdirs==1.4.4
bcrypt==3.2.0
cffi==1.14.4
click==7.1.2
cryptography==3.3.1
distlib==0.3.1
filelock==3.0.12
Flask==1.1.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.8
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.4
opencv-python==4.4.0.46
paramiko==2.7.2
pexpect==4.8.0
Pillow==8.0.1
psutil==5.8.0
ptyprocess==0.7.0
pycparser==2.20
PyNaCl==1.4.0
pytz==2020.5
six==1.15.0
virtualenv==20.2.2
Werkzeug==1.0.1
